examples/crewai/requirements.txt
crewai
python-dotenv
-e ../../sdks/python
